Move DIDT test and symlink
