+++ .github/workflows/didt-tests.yaml
@@ -66,7 +66,7 @@ jobs:
           TT_MM_THROTTLE_PERF=5 pytest tests/didt/test_sdxl_matmul.py::test_sdxl_matmul -k "all" --didt-workload-iterations 100 --determinism-check-interval 1
           TT_MM_THROTTLE_PERF=5 pytest tests/didt/test_sdxl_conv_1280x1280_upsample.py::test_sdxl_conv -k "all" --didt-workload-iterations 100 --determinism-check-interval 1
           pytest tests/didt/test_mla_sdpa.py::test_mla_sdpa -k "all" --didt-workload-iterations 10 --determinism-check-interval 1
-          pytest models/demos/deepseek_v3_d_p/tests/didt/test_deepseek_v3_128k_matmul.py -k "1chips or galaxy" --didt-workload-iterations 10 --determinism-check-interval 1
+          pytest tests/didt/test_deepseek_v3_128k_matmul.py -k "1chips or galaxy" --didt-workload-iterations 10 --determinism-check-interval 1
       - uses: tenstorrent/tt-metal/.github/actions/slack-report@main
         if: ${{ failure() }}
         with:
